requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.30
PyMySQL>=1.1.0
pydantic[email]>=2.7.0
email-validator>=2.1.0.post1
pydantic-settings>=2.3.0
alembic>=1.13.2
python-dotenv>=1.0.1
passlib[bcrypt]>=1.7.4
PyJWT>=2.9.0
bcrypt==3.2.2
qrcode[pil]>=7.4.2
aiofiles>=23.2.1
redis>=5.0.0


